First commit with game code
I've placed all of my code on this repository. The repo is named Game,
since I lacked the imagination to come up with a more descriptive name.
I've also modified the gitignore file to ignore the Library folder, per
Unity's advice. Also, be sure to turn on Meta files in Edit->Project
Settings->Editor, in Unity.