SnakeMap now generates rooms randomly
The code in snakemap is terrible, but it seems to be working. It will
attempt to generate 6 random rooms, if they fit. The rooms will always
open off of the East to West hallways, and will always run north or
south. Also, I modified the quest object placement code so that it will
first try placing the objects in a random room. If no rooms are
available, it will stick the quest objects in a random hallway space.