I added basic item support.
I added the Items QuestItem scripts and the necessary resource files to
support them - ItemData/Items.txt and materials and textures in the
Resources directory. It doesn't do much yet, but when a map is
generated, the quest items are placed in the map. When a new map is
generated, all leftover quest items are destroyed and new ones created
for the new map.