Replaced Getters with methods
It turned out that I wasn't able to use Pragma Strict because Unity
Script is unable to determine the type of a value coming from a getter
in an interface. The fix was to either cast the value returned by the
getter, or replace it with a function. Unfortunately, Unity Script
doesn't allow you to cast value types, so I had to convert all of my
getters to functions. I was able to add pragmas strict back to all of
the files. Addresses #12